Add script for the creation of examples archive for releases.
Python script creates an archive of the requried files for standalone running (without building). This does mean that downloading the archive and repository includes duplicates, but this could be resolved with larger restructuring, probably best left for v2.

It may be worth adding an additional script to build the entire examples suite from the command line.