Modify printf statements for each iteration
When using printf anywhere in the singleIteration() function the
lack of \n after 'Processing Simulation Step' results in line mangling
unless the first print begins with a newline.
The printf('Done') is not really required,
either the next iteration will begin and print a new line,
or the total processing time will be output, so Done doesn't add much.