Add Host based agent creation and associated example.
Allows users to specify the creation of agents on the host, and transfer
to target states on the device.
Utility functions are provided to allocate and deallocate agent
structures and arrays of strucutres, including support for agent variable
arrays.
An associated example is included, demonstrating use of the additional
functionality.

Additionally, a new #define is created for the length of agent variable
arrays.