Fix brute force partitioning out of bounds reads/writes
If the block/grid size generated is not an exact match to the population
  or rather is larger than the maximum buffer size,
then out of bounds global reads/writes could occur for agent functions which iterate non-partitioned messaging.

This is prevented by thread bounds checking on global reads / writes, but as the agent function will still execute default values are loaded.
Agent array variables then presented possible issues, so null pointers are passed to the relevant functions which trigger special cases. This was the simplest change which does not require changing of function prototypes (and therfore functions.c).

Closes #189