Implement agent variable access on the host
Memcpy are handelled transparently by tracking if the host data has been updated or not
  This includes resetting the flag if device data is modified by the host function.

Agent array variables and vector types are both supported

The HostAgentCreation examples has been updated to demonstrate this functionality using a custom output function.
  i.e. a CSV file containing a subset of the agent's data is output per iteration in the step function.

Additionally, FLAME GPU keeps track of the iteration number (on the host only), which was required for transparent implementation of memcpy.

It will be worth noting in the documentation that it would be possible to severly hurt performance using these methods.

Closes #178