Implement Agent ID generation on the host and the device
This requires an agent variable named 'id' of an integer type (signed or unsigned) to generate the appropraite function

If an initial states file is provided, the first id assigned by the generate_<agent>_id() function is the maximum+1.
The user can specify a start number if desired.
The value is synchronised between the host and device where appropriate, through memcpy when appropriate.
Device concurrency is handeled using atomicAdd